deepfake/requirements.txt
gradio
torch
facenet-pytorch
numpy
Pillow
opencv-python
pytorch-gradcam
fastapi
websockets
uvicorn
grad-cam